requirements.txt
python-telegram-bot[job-queue]>=21.0
python-dotenv>=1.0.0
# БД: для PostgreSQL в Yandex Cloud (Managed Service)
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
# AI-викторины: Groq (приоритет, бесплатный тир) + Gemini (fallback), извлечение текста
groq>=0.4.0
google-generativeai>=0.8.0
pypdf>=4.0
python-docx>=1.0
httpx>=0.27
